Adds in new notebooks for 2023-10-09 session
This adds a few things:

1. a reference notebook -- which should just be used for reference purposes on Hamilton.
2. a ml one that is split into data/feature engineering and then model fitting.
3. a notebook that shows an LLM workflow doing PDF summarization.

The points of the notebooks are to:

1. Help one understand Hamilton.
2. How one might think about constructing a solution.
3. There is NO CODE to fill in. As long as you read the code, and sequentially execute it -- it should help tell a story.

Note: the notebooks are self-contained, meaning they "write out" modules. In real life you'd then go and take those modules/package them appropriately. But here, I thought it would be easier to show progression by having it all in the notebook and using the order to tell a story.